event tables, seeds, and scheduler created/updated
